+++ autodocodec-yaml/default.nix
@@ -1,14 +1,15 @@
-{ mkDerivation, autodocodec, autodocodec-schema, base, bytestring
-, containers, lib, path, path-io, safe-coloured-text, scientific
-, text, vector, yaml
+{ mkDerivation, autodocodec, autodocodec-exact, autodocodec-schema
+, base, bytestring, containers, lib, path, path-io
+, safe-coloured-text, scientific, text, vector, yaml
 }:
 mkDerivation {
   pname = "autodocodec-yaml";
-  version = "0.4.0.2";
+  version = "0.5.0.0";
   src = ./.;
   libraryHaskellDepends = [
-    autodocodec autodocodec-schema base bytestring containers path
-    path-io safe-coloured-text scientific text vector yaml
+    autodocodec autodocodec-exact autodocodec-schema base bytestring
+    containers path path-io safe-coloured-text scientific text vector
+    yaml
   ];
   homepage = "https://github.com/NorfairKing/autodocodec#readme";
   description = "Autodocodec interpreters for yaml";